Wrote a C++ code to collect two numbers from users, sum the numbers, and output a message of the sum.
This C++ code collects two integers values, n1 and n2 respectively from user and add it together. It then output the message ”The sum of <n1> and <n2> is <n1+n2>. For instance, if user input 4 as n1 and 7 as n2, the message will be "The sum of 4 and 7 is 11".